chore: Address review comments
